feat: add setting to always display typing indicator
Add chatTypingIndicator option to show typing indicator
even when chat is closed. Extracted typing indicator logic
into useTypingIndicatorText hook to avoid duplication.
